fix: resolve tsc errors and drop @ts-nocheck from 11 files
Dropped @ts-nocheck from (no longer needed):
  src/lib/core/eventBus.ts              — contravariance → Set<Function>
  src/lib/services/location/weather.ts   — null guard on format()
  src/lib/services/notifications/guard.ts — iterator on unknown → type assert
  src/lib/services/monitoring/monitors.ts — AstalWl type stubs, removed dead field
  src/widget/applauncher/appButton.tsx   — Object→Widget cast for child
  src/widget/bar/clock.tsx               — optional chain on nullable t
  src/widget/bar/indicators/audio.tsx     — snake_case property cast
  src/widget/bar/indicators/power.tsx     — connectFor widens to Connectable
  src/widget/bar/indicators/recording.tsx — as any binding cast
  src/widget/quicksettings/button-grid/powerprofiles.tsx — same
  src/widget/quicksettings/expander/worldClock.tsx — optional chain

Infrastructure:
  src/lib/core/connectFor.ts             — Connectable interface (GObject | PowerProfiles)
  src/types/gi-modules.d.ts              — type stubs for missing GI modules
                                           (Secret, AstalBrightness, AstalGreet, AstalWl)
  src/types/gnim-overrides.d.ts          — add signal() decorator override for TS6
  src/lib/services/clipboard/history.ts  — re-export ClipboardEntry
  src/lib/services/clipboard/cryptoEngineGhash.ts — Uint8Array generics cast
  src/widget/screenshot-ui/formatQualitySelector.tsx — recordingQuality→ss.prefs
  src/widget/quicksettings/network/apRow.tsx — Promise cast for async method

21 files retain @ts-nocheck for pre-existing GI type gaps (upstream ts-for-gir).

+++ src/lib/services/clipboard/cryptoEngineGhash.ts
@@ -36,7 +36,7 @@ function ghashMul(x: Uint8Array, y: Uint8Array): Uint8Array {
 export function ghash(h: Uint8Array, data: Uint8Array): Uint8Array {
     let y = new Uint8Array(16);
     for (let i = 0; i < data.length; i += 16) {
-        y = ghashMul(xor(y, data.subarray(i, i + 16)), h);
+        y = ghashMul(xor(y, data.subarray(i, i + 16)), h) as Uint8Array;
     }
     return y;
 }
